requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.3
cmake==3.24.0
colorama==0.4.5
comm==0.2.1
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
dlib @ file:///C:/Users/rahul/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl
exceptiongroup==1.2.0
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==24.3.25
fonttools==4.41.1
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imbalanced-learn==0.11.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.3.3
kiwisolver==1.4.4
libclang==18.1.1
lxml==4.9.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
Pillow==9.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pywin32==306
pyzmq==25.1.2
requests==2.32.3
rich==13.7.1
scikit-learn==1.3.1
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
